Update publish-to-azure-webapp-using-vs.md
The document is little vague with regards publishing a site that uses appsettings.json. This lead me to raise a SO query which you guys kindly answered:

https://stackoverflow.com/questions/44768865/asp-net-core-latest-appsettings-json-connectionstring-not-overriding-on-azur/44771419#44771419

I think the document needs to be clearer on how this option effects the use of appsettings.json. I had to manually remove the Config Settings on Azure post-publish (which I wasnt aware of).